Adds test functionality wtih testFunction
Because direct comparison of arrays is not possible, and because tail may return an array with a length greater than 1, an additional function for testing was created (testFunction). This function runs a loop over the expected array and the array that tail outputs, and then calls the assertEqual function on each pair of values extracted by the loop from their respective arrays.